testing/web-platform/tests/fetch/metadata/xslt.tentative.https.sub.html

<!DOCTYPE html>

<link rel="author" href="mtrzos@google.com" title="Maciek Trzos">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=/fetch/metadata/resources/helper.js></script>
<script src=/common/utils.js></script>
<script>
  // Open a window with XML document which loads resources via <?xml-stylesheet/> tag
  let nonce = token();
  let w = window.open("resources/xslt-test.sub.xml?token=" + nonce);
  window.addEventListener('message', function(e) {
    if (e.source != w)
      return;

    // Only testing same-origin XSLT because same-site and cross-site XSLT is blocked.
    promise_test(t => {
      let expected = {"site":"same-origin", "user":"", "mode": "same-origin"};
      return fetch("/fetch/metadata/resources/record-header.py?retrieve=true&file=xslt-same-origin" + nonce)
          .then(response => response.text())
          .then(text => assert_header_equals(text, expected));
    }, "Same-Origin xslt");

    w.close();
  });

</script>
